pyproject.toml
[project]
name = "mcp-ollama-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9.2",
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
]
